+++ src/layouts/default/Default.vue
@@ -1,4 +1,44 @@
 <template>
+  <v-app-bar height="72">
+    <v-toolbar-title class=""><img src="/assets/img/logo.png" alt="Riftlink" /></v-toolbar-title>
+
+    <v-btn>
+      <template v-slot:prepend>
+        <v-icon>mdi-list-box</v-icon>
+      </template>
+      <router-link to="/">Clubes</router-link>
+    </v-btn>
+
+    <v-btn v-if="isAuthenticated">
+      <template v-slot:prepend>
+        <v-icon>mdi-account</v-icon>
+      </template>
+      <router-link to="/profile">Mi perfil</router-link>
+    </v-btn>
+
+    <v-btn @click="login" v-if="!isAuthenticated">
+      <template v-slot:prepend>
+        <v-icon>mdi-login</v-icon>
+      </template>
+      Login
+    </v-btn>
+
+    <v-btn @click="logout" v-if="isAuthenticated">
+      <template v-slot:prepend>
+        <v-icon>mdi-logout</v-icon>
+      </template>
+      Logout
+    </v-btn>
+
+    <v-avatar size="48" class="mr-2" v-if="isAuthenticated">
+      <v-img
+        alt="Avatar"
+        :src="user.picture"></v-img>
+    </v-avatar>
+
+  </v-app-bar>
+
+  <!--
   <v-navigation-drawer floating permanent v-model="drawer">
     <v-list>
       <v-list-item>
@@ -8,7 +48,7 @@
         <template v-slot:prepend>
           <v-icon>mdi-list-box</v-icon>
         </template>
-        <router-link to="/"><v-list-item-title>Ofertas</v-list-item-title></router-link>
+        <router-link to="/"><v-list-item-title>Clubes</v-list-item-title></router-link>
       </v-list-item>
       <v-list-item v-if="isAuthenticated">
         <template v-slot:prepend>
@@ -52,6 +92,7 @@
     </template>
 
   </v-navigation-drawer>
+  -->
 
   <v-main>
     <default-view />