+++ src/views/AdminCreateTeam.vue
@@ -5,29 +5,22 @@
         <!-- Back Button -->
         <v-row>
           <v-col cols="12" sm="8" md="6">
-            <router-link to="/admin/offers"><i class="mdi mdi-arrow-left"></i> Atrás</router-link>
+            <router-link to="/admin/teams"><i class="mdi mdi-arrow-left"></i> Atrás</router-link>
           </v-col>
         </v-row>
 
         <v-card class="mt-3">
           <v-card-title>
             <div>
-              <span class="headline">Crear nueva oferta</span>
+              <span class="headline">Crear nuevo club</span>
             </div>
           </v-card-title>
 
           <v-col cols="12" sm="12" md="12">
-            <!-- Position Name Field -->
-            <v-text-field
-              v-model="offer.positionName"
-              label="Posición"
-              outlined
-            ></v-text-field>
-
             <!-- Team Name Field -->
             <v-text-field
-              v-model="offer.teamName"
-              label="Nombre del equipo"
+              v-model="team.teamName"
+              label="Nombre del club"
               outlined
             ></v-text-field>
 
@@ -41,56 +34,35 @@
                 >
                   <v-img
                     alt="Avatar"
-                    :src="offer.logoUrl"></v-img>
+                    :src="team.logoUrl"></v-img>
                 </v-avatar>
               </v-col>
               <v-col cols="12" md="10">
                 <v-text-field
-                  v-model="offer.logoUrl"
+                  v-model="team.logoUrl"
                   label="URL del avatar"
                   outlined
                 ></v-text-field>
               </v-col>
             </v-row>
 
-            <!-- Rank Field -->
-            <v-text-field
-              v-model="offer.rank"
-              label="Rango requerido"
-              outlined
-            ></v-text-field>
-
-            <!-- Requirements Field -->
-            <v-textarea
-              v-model="offer.requirements"
-              label="Qué pedimos"
-              outlined
-            ></v-textarea>
-
             <!-- About Us Field -->
             <v-textarea
-              v-model="offer.aboutUs"
+              v-model="team.aboutUs"
               label="Qué ofrecemos"
               outlined
             ></v-textarea>
 
             <!-- Apply Contact Field -->
             <v-text-field
-              v-model="offer.applyContact"
+              v-model="team.applyContact"
               label="Usuario de Discord de contacto"
-              outlined
-            ></v-text-field>
-
-            <!-- Apply Url Field -->
-            <v-text-field
-              v-model="offer.applyUrl"
-              label="Link para aplicar"
               outlined
             ></v-text-field>
 
             <!-- Checkbox Active -->
             <v-checkbox
-                v-model="offer.active"
+                v-model="team.active"
                 label="Activa"></v-checkbox>
           </v-col>
         </v-card>
@@ -105,28 +77,28 @@
           <v-col cols="12" sm="12" md="12">
             <!-- Apply Contact Field -->
             <v-text-field
-              v-model="offer.websiteLink"
+              v-model="team.websiteLink"
               label="Sitio web"
               outlined
             ></v-text-field>
 
             <!-- Apply Contact Field -->
             <v-text-field
-              v-model="offer.linktreeHandle"
+              v-model="team.linktreeHandle"
               label="Linktree handle"
               outlined
             ></v-text-field>
 
             <!-- Apply Contact Field -->
             <v-text-field
-              v-model="offer.discordInvite"
+              v-model="team.discordInvite"
               label="Discord invite"
               outlined
             ></v-text-field>
 
             <!-- Apply Contact Field -->
             <v-text-field
-              v-model="offer.twitterHandle"
+              v-model="team.twitterHandle"
               label="Twitter handle"
               outlined
             ></v-text-field>
@@ -138,7 +110,7 @@
     <!-- Create Button -->
     <v-row justify="center">
       <v-col cols="12" sm="8" md="6">
-        <v-btn block color="primary" @click="createOffer">Crear</v-btn>
+        <v-btn block color="primary" @click="createTeam">Crear</v-btn>
       </v-col>
     </v-row>
 
@@ -152,7 +124,7 @@ import Alert from '@/components/Alert.vue';
 
 import { useAuth0 } from '@auth0/auth0-vue';
 
-import offersApiClient from "@/services/OffersApiClient.js"
+import teamsApiClient from "@/services/TeamsApiClient.js"
 
 export default {
   components: {
@@ -161,13 +133,9 @@ export default {
   data() {
     return {
       loading: true,
-      offer: {
+      team: {
         teamName: '',
-        positionName: '',
-        rank: '',
         logoUrl: '',
-        applyUrl: '',
-        requirements: '',
         aboutUs: '',
         applyContact: '',
         websiteLink: '',
@@ -186,14 +154,14 @@ export default {
     }
   },
   methods: {
-    async createOffer() {
+    async createTeam() {
       try {
         const accessToken = await this.getAccessTokenSilently()
-        await offersApiClient.createOffer(accessToken, this.offer)
-        this.$router.push('/admin/offers');
+        await teamsApiClient.createTeam(accessToken, this.team)
+        this.$router.push('/admin/teams');
       } catch (error) {
-        this.$refs.alert.alertError("¡Ups! Hubo un problema al crear la oferta.")
-        console.error("Couldn't save offer:", error);
+        this.$refs.alert.alertError("¡Ups! Hubo un problema al crear el club.")
+        console.error("Couldn't save team:", error);
       }
     },
   },