+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Offers are now teams.
+- Navigation drawer is now a navbar.
+
+### Removed
+
+- All offer features and information.
+
 ## [1.2.0] - 2023-11-01
 
 ### Added